requirements.txt
psycopg2-binary
python-dotenv
httpx
beautifulsoup4
pydantic
